Moved Plane into Geometry namespace
